// packages/core/src/payload-types.ts
export interface Server {
// ...
  dflowVpsDetails?: {
    orderId?: string | null;
    instanceId?: number | null;
    status?:
      | (
          | 'provisioning'
          | 'uninstalled'
          | 'running'
          | 'stopped'
          | 'error'
          | 'installing'
          | 'unknown'
          | 'manual_provisioning'
          | 'product_not_available'
          | 'verification_required'
          | 'rescue'
          | 'pending_payment'
          | 'other'
          | 'reset_password'
        )
      | null;
    next_billing_date?: string | null;
  };
// ...
}